@@ -14,8 +14,8 @@
     "axios": "^0.26.0",
     "font-awesome": "^4.7.0",
     "html2canvas": "^1.4.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
     "react-share": "^4.4.0",
